Simplify test
